spam feature is now author dependent

+++ src/chocobot/app.py
@@ -28,7 +28,7 @@ async def on_message(message: discord.Message) -> None:
 
     # Prevent spam (2 or more identical messages in a row)
     messages: list[discord.Message] = [msg async for msg in message.channel.history(limit=2)]
-    if all(msg.content == message.content for msg in messages):
+    if all(msg.content == message.content and msg.author == message.author for msg in messages):
         await message.delete()
         await message.channel.send(f"{message.author.mention}, please don't spam!")
